Refactor render module: remove classes, make render_config module-level function
- Move render_template.py to hook_templates/render.py
- Convert TemplateRenderer class to module-level functions
- Make render_config the only public function
- Update imports and exports in __init__.py
- Fix template paths and test expectations
- Remove unnecessary main function and docstrings
