+++ pre_commit_starter/hook_templates/__init__.py
@@ -1,1 +1,23 @@
 """Hook generation modules."""
+
+from .render import (
+    render_config,
+    generate_base_hooks,
+    generate_python_hooks,
+    generate_docker_hooks,
+    generate_js_hooks,
+    generate_go_hooks,
+    generate_github_actions_hooks,
+    generate_meta_wrapper,
+)
+
+__all__ = [
+    "render_config",
+    "generate_base_hooks",
+    "generate_python_hooks",
+    "generate_docker_hooks",
+    "generate_js_hooks",
+    "generate_go_hooks",
+    "generate_github_actions_hooks",
+    "generate_meta_wrapper",
+]
