+++ pre_commit_starter/hook_templates/python.j2
@@ -7,9 +7,10 @@
   hooks:
   - id: ruff-format
     name: Format Python code with Ruff
+    args: [--line-length=120]
   - id: ruff
     name: Run Ruff linter and formatter
-    args: [--fix]
+    args: [--line-length=120]
 
 - repo: https://github.com/pre-commit/mirrors-mypy
   rev: v1.16.1
