+++ tests/test_render_template_edge_cases.py
@@ -1,7 +1,7 @@
 """Tests for untested edge cases in render_template.py module."""
 
 from pre_commit_starter.config import PreCommitConfig
-from pre_commit_starter.render_template import render_config
+from pre_commit_starter.hook_templates import render_config
 
 
 class TestComplexConfigurations:
